services/qa-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
pymongo==4.6.1
openai==1.10.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0
